+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.2.3]
+ * Added onGlobalStackChanged listener to track page and tab changes
+ ## [0.2.2]
+ * Fixed exception when isFirst isn't set
 ## [0.2.1] 
  * Missing changelog
  * Formatting